Should be handled better?
